requirements.txt
# Core dependencies
streamlit>=1.30.0
streamlit-autorefresh>=1.0.1
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-core>=0.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Web scraping and RSS
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# Finance and data
yfinance>=0.2.0
pandas>=2.0.0

# Search and tools
tavily-python>=0.3.0

# Additional tools
langgraph>=0.0.30
langsmith>=0.0.40

# Utility
pydantic>=2.0.0
typing-extensions>=4.8.0

# Development dependencies (for CI/CD)
pytest>=7.0.0
pytest-mock>=3.12.0
black>=23.0.0
ruff>=0.1.0 
langchain_community